Move MessageRefAdapter to protocol.common package
