+++ src/main/java/com/streamr/client/protocol/common/MessageRefAdapter.java
@@ -1,9 +1,8 @@
-package com.streamr.client.protocol.message_layer;
+package com.streamr.client.protocol.common;
 
 import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonWriter;
-import com.streamr.client.protocol.common.MessageRef;
 import java.io.IOException;
 
 public class MessageRefAdapter extends JsonAdapter<MessageRef> {
