+++ src/main/java/com/streamr/client/protocol/message_layer/StreamMessageV31Adapter.java
@@ -5,6 +5,7 @@ import com.squareup.moshi.JsonDataException;
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonWriter;
 import com.streamr.client.protocol.common.MessageRef;
+import com.streamr.client.protocol.common.MessageRefAdapter;
 import com.streamr.client.protocol.message_layer.StreamMessage.EncryptionType;
 import com.streamr.client.protocol.message_layer.StreamMessage.SignatureType;
 import java.io.IOException;
