+++ src/main/java/com/streamr/client/protocol/message_layer/StreamMessageV32Adapter.java
@@ -4,6 +4,7 @@ import com.squareup.moshi.JsonAdapter;
 import com.squareup.moshi.JsonReader;
 import com.squareup.moshi.JsonWriter;
 import com.streamr.client.protocol.common.MessageRef;
+import com.streamr.client.protocol.common.MessageRefAdapter;
 import com.streamr.client.protocol.message_layer.StreamMessage.EncryptionType;
 import com.streamr.client.protocol.message_layer.StreamMessage.SignatureType;
 import com.streamr.client.utils.EncryptedGroupKey;
